Make sqlite.connection module for ConnectionPlus and move it there
